requirements.txt
-i https://pypi.org/simple
asgiref==3.6.0 ; python_version >= '3.7'
django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
pyjwt==2.6.0 ; python_version >= '3.7'
pytz==2022.7.1
sqlparse==0.4.3 ; python_version >= '3.5'
